Add more variants to macOS testing
In addition to originally testing clang and appleclang via makefiles and
ninja files, need to be able to test via Xcode projects, as an issue has
cropped up there.

However, using regular clang with Xcode installed seems to be very
difficult.

So a second runner has been added and tags added, 'xcode' and 'xclt' for
xcode-command-line-tools, and jobs added for both.